Fix DataSelector performance (#1150)
