+++ packages/react-ui/src/app/features/builder/data-selector/data-selector-utils.ts
@@ -1,14 +1,12 @@
 import { formatUtils } from '@/app/lib/utils';
 import {
   Action,
-  flowHelper,
   isEmpty,
   isNil,
   StepOutputWithData,
   StepWithIndex,
   Trigger,
 } from '@openops/shared';
-import { BuilderState } from '../builder-types';
 
 export type MentionTreeNode = {
   key: string;
@@ -272,26 +270,10 @@ function filterBy(arr: MentionTreeNode[], query: string): MentionTreeNode[] {
   }, [] as MentionTreeNode[]);
 }
 
-/**
- * Selector that computes the path to the target step using flowHelper.findPathToStep
- */
-const getPathToTargetStep = (state: BuilderState) => {
-  const { selectedStep, flowVersion } = state;
-  if (!selectedStep || !flowVersion?.trigger) {
-    return [];
-  }
-  const pathToTargetStep = flowHelper.findPathToStep({
-    targetStepName: selectedStep,
-    trigger: flowVersion.trigger,
-  });
-  return pathToTargetStep;
-};
-
 export const dataSelectorUtils = {
   traverseStepOutputAndReturnMentionTree,
   getAllStepsMentions,
   createTestNode,
   filterBy,
-  getPathToTargetStep,
   hasStepSampleData,
 };
