+++ packages/react-ui/src/app/features/builder/data-selector/tests/data-selector-utils.test.ts
@@ -1,5 +1,4 @@
 import { ActionType, StepWithIndex, TriggerType } from '@openops/shared';
-import { BuilderState } from '../../builder-types';
 import { dataSelectorUtils, MentionTreeNode } from '../data-selector-utils';
 
 jest.mock('@/app/lib/utils', () => ({
@@ -7,47 +6,6 @@ jest.mock('@/app/lib/utils', () => ({
     formatStepInputOrOutput: (v: unknown) => v,
   },
 }));
-
-const createTestTrigger = (name = 'trigger'): any => ({
-  name,
-  displayName: 'Test Trigger',
-  type: TriggerType.BLOCK,
-  settings: {},
-  valid: true,
-  nextAction: null,
-});
-
-const createTestAction = (
-  name: string,
-  displayName: string,
-  nextAction = null,
-): any => ({
-  id: name,
-  name,
-  displayName,
-  type: ActionType.BLOCK,
-  settings: {},
-  valid: true,
-  nextAction,
-});
-
-const createTestFlow = () => {
-  const step3 = createTestAction('step3', 'Step 3');
-  const step2 = createTestAction('step2', 'Step 2', step3);
-  const step1 = createTestAction('step1', 'Step 1', step2);
-  const trigger = createTestTrigger();
-  trigger.nextAction = step1;
-  return trigger;
-};
-
-const createBuilderState = (overrides = {}): BuilderState => {
-  const defaultState: any = {
-    flow: { id: 'flow1', name: 'Test Flow' },
-    flowVersion: { id: 'version1' },
-  };
-
-  return { ...defaultState, ...overrides };
-};
 
 describe('dataSelectorUtils', () => {
   describe('getAllStepsMentions', () => {
@@ -514,53 +472,6 @@ describe('dataSelectorUtils', () => {
 
       const result = dataSelectorUtils.filterBy(nodes, 'match');
       expect(result).toHaveLength(0);
-    });
-  });
-
-  describe('getPathToTargetStep', () => {
-    it.each([
-      {
-        desc: 'returns empty array when selectedStep is not provided',
-        state: createBuilderState({
-          selectedStep: '',
-          flowVersion: { trigger: createTestTrigger() },
-        }),
-      },
-      {
-        desc: 'returns empty array when flowVersion.trigger is not provided',
-        state: createBuilderState({
-          selectedStep: 'step1',
-          flowVersion: {},
-        }),
-      },
-    ])('$desc', ({ state }) => {
-      const result = dataSelectorUtils.getPathToTargetStep(state);
-      expect(result).toEqual([]);
-    });
-
-    it('returns path to the target step', () => {
-      const trigger = createTestFlow();
-      const state = createBuilderState({
-        selectedStep: 'step3',
-        flowVersion: { trigger },
-      });
-
-      const result = dataSelectorUtils.getPathToTargetStep(state);
-      expect(result).toHaveLength(3);
-      expect(result[0].name).toBe('trigger');
-      expect(result[1].name).toBe('step1');
-      expect(result[2].name).toBe('step2');
-    });
-
-    it('returns empty array when target step does not exist', () => {
-      const trigger = createTestFlow();
-      const state = createBuilderState({
-        selectedStep: 'nonExistentStep',
-        flowVersion: { trigger },
-      });
-
-      const result = dataSelectorUtils.getPathToTargetStep(state);
-      expect(result).toEqual([]);
     });
   });
